+++ src/egeo-demo/barrels.ts
@@ -17,6 +17,7 @@ import { StDropdownMenuDemoModule } from './st-dropdown-menu-demo/st-dropdown-me
 import { StFormDemoModule } from './st-form-demo/st-form-demo.module';
 import { StHeaderDemoModule } from './st-header-demo/st-header-demo.module';
 import { StHelpDemoModule } from './st-help-demo/st-help-demo.module';
+import { StHorizontalTabsDemoModule } from './st-horizontal-tabs-demo/st-horizontal-tabs-demo.module';
 import { StInfoBoxDemoModule } from './st-info-box-demo/st-info-box-demo.module';
 import { StInputDemoModule } from './st-input-demo/st-input-demo.module';
 import { StItemListDemoModule } from './st-item-list-demo/st-item-list-demo.module';
@@ -46,6 +47,7 @@ export const EGEO_DEMO_MODULES: any[] = [
    StFormDemoModule,
    StHeaderDemoModule,
    StHelpDemoModule,
+   StHorizontalTabsDemoModule,
    StInfoBoxDemoModule,
    StInputDemoModule,
    StItemListDemoModule,