+++ CHANGELOG.md
@@ -19,7 +19,7 @@
 
 **Breaking changes:**
 
-* st-header: Remove stHeaderBehavior directive because now it's unnecesary
+* st-header: Remove stHeaderBehavior directive because now it's unnecessary
 * st-header: Change contentChangeOffset parameter by changeHeight
 * st-header: Removed app name, now use a ng-content to define
 * st-header: Removed user menu, now use a ng-content to define
@@ -29,7 +29,8 @@
 * st-header: Change general behaviour and design
 * st-input-adjustable: Rename directive 'StInputAdjustable' to st-input-adjustable
 * st-switch: Remove labelPosition input, label always are dispayed at the left.
-
+* st-horizontal-tabs: Event emitted when active tabs changes, now sends the option of type StHorizontalTab 
+* st-horizontal-tabs: Removed functionality to display disabled tabs
 
 **Refactor**
 
@@ -37,6 +38,7 @@
 * st-table: Apply new style according to UX specifications
 * st-switch: Apply new style.
 * st-breadcrumb: Apply new style according to UX specifications
+* st-horizontal-tabs: Apply new style according to UX specifications
 
 **Documentation:**
 
@@ -44,6 +46,7 @@
 * st-checkbox: Added demo
 * st-radio: Added demo
 * st-table: Added demo
+* st-horizontal-tabs: Added demo
 
 **Others**